# Docs/02_contracts_and_schemas/offisim_openapi.yaml
openapi: 3.1.0
info:
  title: Offisim Platform and Local Install APIs
  version: 0.2.0
  summary: Platform marketplace APIs plus local runtime install/import contract snapshots.
  description: >
    This file tracks the current public platform HTTP surface mounted from `apps/platform/src/index.ts`
    together with the local install transaction contract used by `packages/install-core`.
    Public marketplace reads are unauthenticated. Creator publishing, reviews, creator bootstrap,
    install receipt recording, and user library access require authenticated user context.
    Better Auth native session endpoints under `/api/auth/*` and API token management under
    `/v1/auth/tokens` are implementation-owned and intentionally not expanded here.
    Local install/import endpoints remain declaration-driven: no arbitrary install hooks,
    no postinstall scripts, and no remote code execution during import.
servers:
  - url: https://api.offisim.market
    description: Registry and marketplace API
  - url: http://127.0.0.1:43111
    description: Local desktop runtime API
...
security: []
paths:
...
  /v1/market/listings/{listingId}/reviews:
    get:
      tags: [Market Public, Reviews]
      summary: List visible reviews for a listing
      operationId: listListingReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                type: object
                required: [listing_id, reviews]
                properties:
                  listing_id: { type: string }
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
...
  /v1/publish/me:
    get:
      tags: [Creator Publish]
      summary: Get the authenticated user's creator profile, if one exists
      operationId: getMyCreatorProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Creator profile or null
          content:
            application/json:
              schema:
                type: object
                required: [creator]
                properties:
                  creator:
                    anyOf:
                      - $ref: '#/components/schemas/CreatorDetail'
                      - type: 'null'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/publish/drafts:
    get:
      tags: [Creator Publish]
      summary: List publish drafts for the authenticated creator
      operationId: listPublishDrafts
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, validated, submitted, approved, rejected]
      responses:
        '200':
          description: Draft list
          content:
            application/json:
              schema:
                type: object
                required: [drafts]
                properties:
                  drafts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublishDraft'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Creator Publish]
      summary: Create a publish draft
      operationId: createPublishDraft
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftRequest'
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishDraft'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/publish/drafts/{draftId}:
    get:
      tags: [Creator Publish]
      summary: Get draft detail
      operationId: getPublishDraft
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Draft detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishDraft'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Creator Publish]
      summary: Delete a non-submitted draft
      operationId: deletePublishDraft
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Draft deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted, draft_id]
                properties:
                  deleted: { type: boolean }
                  draft_id: { type: string }
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/publish/drafts/{draftId}/manifest:
    put:
      tags: [Creator Publish]
      summary: Attach or replace draft manifest metadata
      operationId: putDraftManifest
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutDraftManifestRequest'
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishDraft'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/publish/submit:
    post:
      tags: [Creator Publish]
      summary: Submit a validated draft for review and listing publication
      operationId: submitPublishDraft
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishSubmitRequest'
      responses:
        '202':
          description: Draft accepted for moderation workflow
          content:
            application/json:
              schema:
                type: object
                required: [draft_id, moderation_job_id, status]
                properties:
                  draft_id: { type: string }
                  moderation_job_id: { type: string }
                  status:
                    type: string
                    enum: [queued, pending_review]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user is not a creator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /v1/reviews:
    post:
      tags: [Reviews]
      summary: Create or replace a review for a listing
      operationId: upsertReview
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Existing review updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '201':
          description: Review written
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token or Better Auth bearer session
  parameters:
    ListingId:
      name: listingId
      in: path
      required: true
      schema: { type: string }
...
    DraftId:
      name: draftId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ListingSummary:
      type: object
      required:
        [listing_id, slug, kind, title, summary, creator, status, latest_version, rating, install_count]
      properties:
        listing_id: { type: string }
        slug: { type: string }
        kind:
          type: string
          enum: [employee, skill, sop, company_template, office_layout, bundle, prefab]
        title: { type: string }
        summary: { type: string }
        creator:
          $ref: '#/components/schemas/CreatorSummary'
        status:
          type: string
          enum: [listed, hidden, retired]
        latest_version: { type: string }
        rating: { type: number, format: float }
        install_count: { type: integer, minimum: 0 }
        tags:
          type: array
          items: { type: string }
        preview:
          $ref: '#/components/schemas/PreviewRef'
...
    CreatorSummary:
      type: object
      required: [creator_id, handle, display_name, verification_state]
      properties:
        creator_id: { type: string }
        handle: { type: string }
        display_name: { type: string }
        verification_state:
          type: string
          enum: [unverified, verified, trusted]
    CreatorDetail:
      allOf:
        - $ref: '#/components/schemas/CreatorSummary'
        - type: object
          required: [created_at, listings]
          properties:
            bio: { type: string, nullable: true }
            website_url: { type: string, nullable: true }
            created_at:
              type: string
              format: date-time
            listings:
              type: array
              items:
                $ref: '#/components/schemas/ListingSummary'
...
    PreviewRef:
      type: object
      properties:
        kind:
          type: string
          enum: [icon, image, video, readme]
        url:
          type: string
          format: uri
        alt: { type: string }
    PublishDraft:
      type: object
      required: [draft_id, creator_id, kind, title, status, created_at, updated_at]
      properties:
        draft_id: { type: string }
        creator_id: { type: string }
        listing_id: { type: string, nullable: true }
        kind:
          type: string
          enum: [employee, skill, sop, company_template, office_layout, bundle, prefab]
        title: { type: string }
        summary: { type: string, nullable: true }
        artifact_id: { type: string, nullable: true }
        manifest_json:
          type: object
          additionalProperties: true
        validation_state:
          type: string
          enum: [unknown, valid, invalid]
        validation_report:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [draft, validated, submitted, approved, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateDraftRequest:
      type: object
      required: [kind, title]
      properties:
        kind:
          type: string
          enum: [employee, skill, sop, company_template, office_layout, bundle, prefab]
        listing_id:
          type: string
          nullable: true
        title:
          type: string
          minLength: 1
        summary:
          type: string
    PutDraftManifestRequest:
      type: object
      required: [manifest_json]
      properties:
        manifest_json:
          type: object
          additionalProperties: true
        artifact:
          type: object
          properties:
            external_url:
              type: string
              format: uri
            sha256:
              type: string
            size_bytes:
              type: integer
              minimum: 0
    PublishSubmitRequest:
      type: object
      required: [draft_id]
      properties:
        draft_id: { type: string }
        submit_message: { type: string }
...
    CreateReviewRequest:
      type: object
      required: [listing_id, rating]
      properties:
        listing_id: { type: string }
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
...
    Review:
      type: object
      required: [review_id, listing_id, rating, moderation_state, created_at, updated_at]
      properties:
        review_id: { type: string }
        listing_id: { type: string }
        user_id: { type: string }
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title: { type: string }
        body: { type: string }
        moderation_state:
          type: string
          enum: [visible, hidden, flagged]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true